refactor: 코드 수정
	1. api로 spot검색은 한번만하고, db에서 검색하는 구조로 바꿈
	2. spotDetail.html 지도, 좋아요 추가
	3. spotList.html 검색 결과 없을 때 화면 이미지 추가
	4. spotList, spotDetail.html 카테고리 표시
	5. 각 category 탭 눌렀을 때 해당 카테고리에 대한 전체 검색 결과 보여주기
	6. spotDetail.html fileUrl로 가져와서 보여주기

+++ src/main/java/com/kvp/thinkjeonju/model/Spot.java
@@ -23,7 +23,7 @@ public class Spot {
 	private Double posY;
 	private String url;
 	private String tel;
-	private int fileCnt;
+	private long fileCnt;
 	private int category;
 	
 	public static Spot from(SpotDTO spotDTO) {
